invoice_ocr/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2

# AI
anthropic>=0.40.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# PDF handling
pdfplumber==0.11.4
pdf2image==1.17.0
Pillow==10.4.0

# OCR
pytesseract==0.3.13

# Structured formats
lxml==5.3.0
pandas==2.2.3
openpyxl==3.1.5

# Email ingestion
# imaplib is stdlib

# Queue
celery==5.4.0
redis==5.1.1

# HTTP
httpx==0.27.2
aiofiles==24.1.0

# Utilities
python-dotenv==1.0.1
jinja2==3.1.4
# hashlib is stdlib

# Dev / testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
